+++ packages/mode/index.ts
@@ -1,5 +1,5 @@
 import { rotate } from "@tonaljs/collection";
-import { deprecate, Named, NoteName, transpose } from "@tonaljs/core";
+import { deprecate, NamedPitch, NoteName, transpose } from "@tonaljs/core";
 import { simplify, transposeFifths } from "@tonaljs/interval";
 import { EmptyPcset, Pcset } from "@tonaljs/pcset";
 import { get as getType } from "@tonaljs/scale-type";
@@ -14,7 +14,7 @@ const MODES = [
   [6, 3434, 5, "locrian", "dim", "m7b5"],
 ] as const;
 
-type ModeDatum = typeof MODES[number];
+type ModeDatum = (typeof MODES)[number];
 
 export interface Mode extends Pcset {
   readonly name: string;
@@ -44,7 +44,7 @@ modes.forEach((mode) => {
   });
 });
 
-type ModeLiteral = string | Named;
+type ModeLiteral = string | NamedPitch;
 
 /**
  * Get a Mode by it's name
@@ -69,8 +69,8 @@ export function get(name: ModeLiteral): Mode {
   return typeof name === "string"
     ? index[name.toLowerCase()] || NoMode
     : name && name.name
-    ? get(name.name)
-    : NoMode;
+      ? get(name.name)
+      : NoMode;
 }
 
 export const mode = deprecate("Mode.mode", "Mode.get", get);
@@ -137,7 +137,7 @@ export function distance(destination: ModeLiteral, source: ModeLiteral) {
 export function relativeTonic(
   destination: ModeLiteral,
   source: ModeLiteral,
-  tonic: NoteName
+  tonic: NoteName,
 ) {
   return transpose(tonic, distance(destination, source));
 }