+++ packages/pitch/test.ts
@@ -1,11 +1,11 @@
 import {
+  Pitch,
   chroma,
-  height,
-  midi,
-  freq,
   coordinates,
+  height,
+  isNamedPitch,
+  midi,
   pitch,
-  IntervalPitch,
 } from "./index";
 
 // Pitch classes
@@ -20,10 +20,13 @@ const A4 = { step: 5, alt: 0, oct: 4 };
 const Gs6 = { step: 4, alt: 1, oct: 6 };
 
 // Intervals
-const P5: IntervalPitch = { step: 4, alt: 0, oct: 0, dir: 1 };
-const P_5: IntervalPitch = { step: 4, alt: 0, oct: 0, dir: -1 };
+const P5: Pitch = { step: 4, alt: 0, oct: 0, dir: 1 };
+const P_5: Pitch = { step: 4, alt: 0, oct: 0, dir: -1 };
 
 describe("@tonaljs/pitch", () => {
+  test("isNamedPitch", () => {
+    expect(isNamedPitch({ name: "C" })).toBe(true);
+  });
   test("height", () => {
     expect([C, Cs, Cb, A].map(height)).toEqual([-1200, -1199, -1201, -1191]);
     expect([C4, A4, Gs6].map(height)).toEqual([48, 57, 80]);