+++ packages/roman-numeral/index.ts
@@ -5,11 +5,11 @@ import {
   interval,
   isNamed,
   isPitch,
-  Named,
   Pitch,
 } from "@tonaljs/core";
 
-export interface RomanNumeral extends Pitch, Named {
+export interface RomanNumeral extends Pitch {
+  readonly name: string;
   readonly empty: boolean;
   readonly roman: string;
   readonly interval: string;
@@ -42,22 +42,23 @@ const cache: Record<string, RomanNumeral | NoRomanNumeral> = {};
  * @example
  * romanNumeral("VIIb5") // => { name: "VII", type: "b5", num: 7, major: true }
  */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function get(src: any): RomanNumeral | NoRomanNumeral {
   return typeof src === "string"
     ? cache[src] || (cache[src] = parse(src))
     : typeof src === "number"
-    ? get(NAMES[src] || "")
-    : isPitch(src)
-    ? fromPitch(src)
-    : isNamed(src)
-    ? get(src.name)
-    : NoRomanNumeral;
+      ? get(NAMES[src] || "")
+      : isPitch(src)
+        ? fromPitch(src)
+        : isNamed(src)
+          ? get(src.name)
+          : NoRomanNumeral;
 }
 
 const romanNumeral = deprecate(
   "RomanNumeral.romanNumeral",
   "RomanNumeral.get",
-  get
+  get,
 );
 
 /**