+++ packages/modules/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tonaljs/modules
 
+## 4.8.2
+
+### Patch Changes
+
+- Named type was renamed to NamedPitch. Add old export for backwards compatibility
+- Updated dependencies
+  - tonal@5.1.3
+
 ## 4.8.1
 
 ### Patch Changes