+++ packages/range/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tonaljs/range
 
+## 4.8.2
+
+### Patch Changes
+
+- Named type was renamed to NamedPitch. Add old export for backwards compatibility
+- Updated dependencies
+  - @tonaljs/collection@4.8.1
+  - @tonaljs/midi@4.9.1
+
 ## 4.8.1
 
 ### Patch Changes