+++ packages/scale-type/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tonaljs/scale-type
 
+## 4.8.3
+
+### Patch Changes
+
+- Named type was renamed to NamedPitch. Add old export for backwards compatibility
+- Updated dependencies
+  - @tonaljs/core@4.10.3
+  - @tonaljs/pcset@4.8.3
+
 ## 4.8.1
 
 ### Patch Changes