+++ packages/scale/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @tonaljs/scale
 
+## 4.12.3
+
+### Patch Changes
+
+- Named type was renamed to NamedPitch. Add old export for backwards compatibility
+- Updated dependencies
+  - @tonaljs/chord-type@5.0.3
+  - @tonaljs/collection@4.8.1
+  - @tonaljs/core@4.10.3
+  - @tonaljs/note@4.10.1
+  - @tonaljs/pcset@4.8.3
+  - @tonaljs/scale-type@4.8.3
+
 ## 4.12.2
 
 ### Patch Changes