+++ packages/voice-leading/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tonaljs/voice-leading
 
+## 5.0.1
+
+### Patch Changes
+
+- Named type was renamed to NamedPitch. Add old export for backwards compatibility
+- Updated dependencies
+  - @tonaljs/note@4.10.1
+
+## 5.0.0
+
+### Major Changes
+
+- 15017c0: First published version
+
 ## 4.8.0
 
 ### Minor Changes