+++ packages/voicing/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @tonaljs/voicing
 
+## 5.0.1
+
+### Patch Changes
+
+- Named type was renamed to NamedPitch. Add old export for backwards compatibility
+- Updated dependencies
+  - @tonaljs/chord@5.0.1
+  - @tonaljs/interval@4.8.1
+  - @tonaljs/note@4.10.1
+  - @tonaljs/range@4.8.2
+  - @tonaljs/voice-leading@5.0.1
+  - @tonaljs/voicing-dictionary@5.0.1
+
+## 5.0.0
+
+### Major Changes
+
+- 15017c0: First published version
+
+### Patch Changes
+
+- Updated dependencies [15017c0]
+  - @tonaljs/voicing-dictionary@5.0.0
+  - @tonaljs/voice-leading@5.0.0
+
 ## 4.8.1
 
 ### Patch Changes